generate_dataset.py: reverse GT binarization to black text on white background
Previously, the generated ground truth (GT) images had black backgrounds
with white text (texture) due to the use of THRESH_BINARY_INV in adaptive
thresholding. This commit removes the inversion, switching to THRESH_BINARY
instead. As a result, the GT now has a white background with black text,
which aligns with the standard format used in other datasets and segmentation
pipelines.

Signed-off-by: Charlie Chiu <charlie910417@gmail.com>